+++ sdk/CHANGELOG.md
@@ -1,5 +1,29 @@
 # Changelog
 
+## [1.2.3](https://github.com/flipboxlabs/aws-audit/compare/sdk-v1.2.2...sdk-v1.2.3) (2026-04-06)
+
+
+### Bug Fixes
+
+* **sdk:** lazy-evaluate process.env.SERVICE in defineAuditConfig ([c12ac4f](https://github.com/flipboxlabs/aws-audit/commit/c12ac4fe4573642eefa3d2f4f8e9e37046b4d9de))
+* **sdk:** replace AES-256-CTR pagination tokens with base64url encoding ([916be50](https://github.com/flipboxlabs/aws-audit/commit/916be50f7467bbe76c38a2f60d3c08dad30ad84c))
+
+## [1.2.2](https://github.com/flipboxlabs/aws-audit/compare/sdk-v1.2.1...sdk-v1.2.2) (2026-04-06)
+
+
+### Bug Fixes
+
+* use Array.length instead of Object.keys for storedAudits size check ([2fdd872](https://github.com/flipboxlabs/aws-audit/commit/2fdd872199a86cea43c6b0c545efe8165f2e6832))
+
+## [1.2.1](https://github.com/flipboxlabs/aws-audit/compare/sdk-v1.2.0...sdk-v1.2.1) (2026-04-06)
+
+
+### Bug Fixes
+
+* format CHANGELOG.md to pass vp check ([cb634de](https://github.com/flipboxlabs/aws-audit/commit/cb634de7fc81caf27a511166204afdfffbd6b7d0))
+* **repository:** retry BatchWriteItem UnprocessedItems with exponential backoff ([7130a17](https://github.com/flipboxlabs/aws-audit/commit/7130a179b83fc7ef6edc800b317d333f7951e587))
+* **repository:** retry BatchWriteItem UnprocessedItems with exponential backoff ([5a95440](https://github.com/flipboxlabs/aws-audit/commit/5a95440a911b7f6559534c9066e75b12995f9ee1)), closes [#18](https://github.com/flipboxlabs/aws-audit/issues/18)
+
 ## [1.2.0](https://github.com/flipboxlabs/aws-audit/compare/sdk-v1.1.3...sdk-v1.2.0) (2026-04-04)